pyproject.toml
[tool.poetry]
name = "fastapi-strawberry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.96.0"}

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
